requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
boto3==1.16.63
botocore==1.19.63
cachetools==4.2.2
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
cycler==0.10.0
decorator==4.4.1
Django==2.0.5
django-cors-headers==2.2.0
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.17.3
django-filter==1.1.0
django-nine==0.1.13
django-webpack-loader==0.6.0
djangorestframework==3.8.2
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
Faker==0.8.8
ffprobe3==0.1.2
flake8==3.5.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==19.8.1
httplib2==0.19.1
idna==2.7
imageio==2.6.1
imageio-ffmpeg==0.3.0
ipaddress==1.0.22
jmespath==0.10.0
kiwisolver==1.1.0
lxml==4.2.5
Markdown==2.6.11
matplotlib==3.0.3
mccabe==0.6.1
mixer==6.0.1
mock==2.0.0
more-itertools==4.2.0
mysql-replication==0.21
mysqlclient==1.3.12
numpy==1.17.4
oauthlib==3.1.0
pbr==5.1.2
pdfkit==0.6.1
Pillow==5.3.0
pkg-resources==0.0.0
pluggy==0.6.0
proglog==0.1.9
protobuf==3.16.0
py==1.5.3
py-bcrypt==0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.3
PyMySQL==0.9.3
pyparsing==2.4.2
pytest==3.6.0
pytest-cov==2.5.1
pytest-django==3.2.1
python-dateutil==2.7.3
python-memcached==1.59
python-pptx==0.6.15
python-resize-image==1.1.18
pytz==2018.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.7
six==1.16.0
SpeechRecognition==3.8.1
text-unidecode==1.2
tqdm==4.40.2
uritemplate==3.0.1
urllib3==1.26.0
Wand==0.5.2
xlrd==1.2.0
XlsxWriter==1.1.1
xlwt==1.3.0
